Convert controller test to request test
This further decouples these tests from the implementation (i.e. which
controller are we calling?).
